Handle read/write more literally
Make sure read-in of SRT handles the case where an extra newline is
present at the end of the final subtitle block (due to it not being
split out correctly). This is achieved by stripping out extra newlines
at the end of the text portion of blocks, since we would never expect
to see this anyway.

Also changes how collapsing blocks is done in the write function.